+++ envs/monkey_zoo/blackbox/test_configurations/depth_3_a.py
@@ -16,7 +16,7 @@ from .utils import (
 
 # Tests:
 #     Powershell (10.2.3.45, 10.2.3.46, 10.2.3.47, 10.2.3.48)
-#     Tunneling (SSH brute force) (10.2.2.9, 10.2.1.10, 10.2.0.12, 10.2.0.11)
+#     Tunneling through grandparent agent (SSH brute force) (10.2.2.9, 10.2.1.10, 10.2.0.11)
 #     WMI pass the hash (10.2.2.15)
 
 
@@ -38,7 +38,6 @@ def _add_subnets(agent_configuration: AgentConfiguration) -> AgentConfiguration:
         "10.2.3.47",
         "10.2.3.48",
         "10.2.1.10",
-        "10.2.0.12",
         "10.2.0.11",
         "10.2.2.15",
     ]
@@ -62,7 +61,6 @@ CREDENTIALS = (
     Credentials(None, Password("Passw0rd!")),
     Credentials(None, Password("3Q=(Ge(+&w]*")),
     Credentials(None, Password("`))jU7L(w}")),
-    Credentials(None, Password("t67TC5ZDmz")),
     Credentials(None, NTHash("d0f0132b308a0c4e5d1029cc06f48692")),
     Credentials(None, NTHash("5da0889ea2081aa79f6852294cba4a5e")),
     Credentials(None, NTHash("50c9987a6bf1ac59398df9f911122c9b")),