+++ monkey/infection_monkey/credential_collectors/ssh_collector/ssh_handler.py
@@ -175,7 +175,7 @@ def _publish_credentials_stolen_event(
     credentials_stolen_event = CredentialsStolenEvent(
         source=get_agent_id(),
         tags=SSH_COLLECTOR_EVENT_TAGS,
-        stolen_credentials=[collected_credentials],
+        stolen_credentials=collected_credentials,
     )
 
     event_queue.publish(credentials_stolen_event)