+++ monkey/infection_monkey/model/victim_host_factory.py
@@ -4,7 +4,6 @@ from typing import Optional, Tuple
 from infection_monkey.model import VictimHost
 from infection_monkey.network import NetworkAddress
 from infection_monkey.network.tools import get_interface_to_target
-from infection_monkey.tunnel import MonkeyTunnel
 
 logger = logging.getLogger(__name__)
 
@@ -12,12 +11,10 @@ logger = logging.getLogger(__name__)
 class VictimHostFactory:
     def __init__(
         self,
-        tunnel: Optional[MonkeyTunnel],
         island_ip: Optional[str],
         island_port: Optional[str],
         on_island: bool,
     ):
-        self.tunnel = tunnel
         self.island_ip = island_ip
         self.island_port = island_port
         self.on_island = on_island
@@ -26,19 +23,15 @@ class VictimHostFactory:
         domain = network_address.domain or ""
         victim_host = VictimHost(network_address.ip, domain)
 
-        if self.tunnel:
-            victim_host.default_tunnel = self.tunnel.get_tunnel_for_ip(victim_host.ip_addr)
-
         if self.island_ip:
             ip, port = self._choose_island_address(victim_host.ip_addr)
             victim_host.set_island_address(ip, port)
 
-        logger.debug(f"Default tunnel for {victim_host} set to {victim_host.default_tunnel}")
         logger.debug(f"Default server for {victim_host} set to {victim_host.default_server}")
 
         return victim_host
 
-    def _choose_island_address(self, victim_ip: str) -> Tuple[str, Optional[str]]:
+    def _choose_island_address(self, victim_ip: str) -> Tuple[Optional[str], Optional[str]]:
         # Victims need to connect back to the interface they can reach
         # On island, choose the right interface to pass to children monkeys
         if self.on_island: